Week 2 Assignment
Q1: Multiplication Table
Write a Python program that displays a multiplication table.
The program should ask the user for the number they want to use for the multiplication table.
Using a loop, display the multiplication table for the number.
End at the number x 12.
Your output should look like below, taking 4 as the number.

Q2: Book Library
Create a dictionary with keys as author names and values as lists of book titles.
Ask the user for the author or book title they're searching for.
Use a for loop to iterate through the dictionary.
If the author name or book title matches the user's input, list the relevant information about the book.
Allow adding new books and authors.
Ask for the author name and book title and update the dictionary with the new entry.
Recommend books based on.
Book titles or author names. Your program should be able to loop through all keys and values in the
dictionary and find a matching item for the user's query. If no entry is found, tell the user there is no
such book in your dictionary.

print("Happy Coding!!!😎")